Changed position to float
It's good practice for position to be float because if in the future we would like to add moving effects that are not quite round numbers

+++ core/src/com/group25/game/Creature.java
@@ -6,7 +6,7 @@ public abstract class Creature extends GameEntity{
 
     private int health;
 
-    public Creature(int positionX, int positionY,int width, int height,int health, Sprite img, int entitySpeed){
+    public Creature(float positionX, float positionY,int width, int height,int health, Sprite img, int entitySpeed){
         super(positionX, positionY, width, height, img, entitySpeed);
     }
